+++ src/js/analytics/ga.js
@@ -1,4 +1,7 @@
 export default function () {
+
+    console.log('Initialize Google Analytics;');
+
     var _gaq = _gaq || [];
     _gaq.push(['_setAccount', 'UA-60598728-1']);
     _gaq.push(['_trackPageview']);
@@ -7,8 +10,7 @@ export default function () {
     ga.type = 'text/javascript';
     ga.async = true;
     ga.src = 'https://ssl.google-analytics.com/ga.js';
+
     var s = document.getElementsByTagName('script')[0];
     s.parentNode.insertBefore(ga, s);
-
-    console.log('GA');
 }
