+++ src/manifest.json
@@ -32,7 +32,7 @@
   "content_security_policy": "script-src 'self' https://ssl.google-analytics.com; object-src 'self'",
   "browser_action": {
     "default_icon": "img/icon48.png",
-    "default_popup": "news.html"
+    "default_popup": "popup.html"
   },
   "icons": {
     "16": "img/icon16.png",
